# swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Pymes web api",
    "description": "web api de la aplicacion Pymes"
  },
  "host": "",
  "basePath": "",
  "tags": [
    {
      "name": "ArticulosFamilias",
      "description": "apis de ArticulosFamilias"
    },
    {
      "name": "Articulos",
      "description": "apis de Articulos"
    },
    {
      "name": "Seguridad",
      "description": "apis de Seguridad"
    },
    {
      "name": "Test Formularios",
      "description": "apis para testing de Action de Formularios"
    }
  ],
  "schemes": [],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "formato: Bearer eyJhbGciOi...vwI"
    }
  },
  "paths": {
    "/api/articulos": {
      "get": {
        "tags": [
          "Articulos"
        ],
        "summary": "obtiene todos los Articulos",
        "description": "",
        "parameters": [
          {
            "name": "Pagina",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Nombre",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Activo",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articulos/{id}": {
      "get": {
        "tags": [
          "Articulos"
        ],
        "summary": "obtiene un Articulo",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del Articulo..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Articulos"
        ],
        "summary": "actualiza un Articulo",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del Articulo..."
          },
          {
            "name": "Articulo",
            "in": "body",
            "description": "Articulo a actualizar",
            "schema": {
              "$ref": "#/definitions/Articulos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Articulos"
        ],
        "summary": "elimina un Articulo",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del Articulo.."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/articulos/": {
      "post": {
        "tags": [
          "Articulos"
        ],
        "summary": "agrega un Articulos",
        "description": "",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "description": "nuevo Articulo",
            "schema": {
              "$ref": "#/definitions/Articulos"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/jwt/articulos": {
      "get": {
        "tags": [
          "Articulos"
        ],
        "summary": "obtiene todos los Articulos, con seguridad JWT, solo para rol: admin (usuario:admin, clave:123)",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth1": []
          }
        ]
      }
    },
    "/api/articulosfamilias": {
      "get": {
        "tags": [
          "ArticulosFamilias"
        ],
        "summary": "obtiene todos los ArticulosFamilias",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articulosfamilias/testerrorasync": {
      "get": {
        "tags": [
          "ArticulosFamilias"
        ],
        "summary": "test error asincrono",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articulosfamilias/{id}": {
      "get": {
        "tags": [
          "ArticulosFamilias"
        ],
        "summary": "obtiene un ArticuloFamilia",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del ArticulosFamilias..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "ArticulosFamilias"
        ],
        "summary": "actualiza un ArticuloFamilia",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del ArticulosFamilias..."
          },
          {
            "name": "ArticulosFamilias",
            "in": "body",
            "description": "ArticulosFamilias a actualizar",
            "schema": {
              "$ref": "#/definitions/ArticulosFamilias"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "ArticulosFamilias"
        ],
        "summary": "elimina un ArticuloFamilia",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del ArticulosFamilias..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articulosfamilias/": {
      "post": {
        "tags": [
          "ArticulosFamilias"
        ],
        "summary": "agrega un ArticuloFamilia",
        "description": "",
        "parameters": [
          {
            "name": "ArticulosFamilias",
            "in": "body",
            "description": "nuevo ArticulosFamilias",
            "schema": {
              "$ref": "#/definitions/ArticulosFamilias"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articulosfamiliasmock": {
      "get": {
        "tags": [
          "ArticulosFamiliasMock"
        ],
        "summary": "obtiene todos los ArticulosFamilias",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/articulosfamiliasmock/{id}": {
      "get": {
        "tags": [
          "ArticulosFamiliasMock"
        ],
        "summary": "obtiene un ArticulosFamiliasMock",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del articulosfamiliasmock..."
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "ArticulosFamiliasMock"
        ],
        "summary": "actualiza un ArticulosFamilias",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del articulosfamiliasmock..."
          },
          {
            "name": "articulosfamiliasmock",
            "in": "body",
            "description": "articulosfamiliasmock a actualizar",
            "schema": {
              "$ref": "#/definitions/ArticulosFamiliasMock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "ArticulosFamiliasMock"
        ],
        "summary": "elimina un ArticulosFamilias",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "identificador del articulosfamiliasmock..."
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/articulosfamiliasmock/": {
      "post": {
        "tags": [
          "ArticulosFamiliasMock"
        ],
        "summary": "agrega un ArticulosFamilias",
        "description": "",
        "parameters": [
          {
            "name": "articulosfamiliasmock",
            "in": "body",
            "description": "nuevo articulosfamiliasmock",
            "schema": {
              "$ref": "#/definitions/ArticulosFamiliasMock"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/eco": {
      "get": {
        "tags": [
          "Test Formularios"
        ],
        "summary": "Devuelve los name:value recibidos en el querystring en forma de tabla",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Test Formularios"
        ],
        "summary": "Devuelve los name:value recibidos en el body en forma de tabla",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/equipos": {
      "get": {
        "tags": [
          "Equipos Qatar"
        ],
        "summary": "Devuelve los equipos y sus bandera desde un array harcodeado",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/_isalive": {
      "get": {
        "tags": [
          "Test Error"
        ],
        "summary": "Verifica que la aplicacion esta ejecutando correctamente!",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/testerrorsincrono": {
      "get": {
        "tags": [
          "Test Error"
        ],
        "summary": "Desencadena un error sincrono",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/testerrorasincrono1": {
      "get": {
        "tags": [
          "Test Error"
        ],
        "summary": "Desencadena un error asincrono: controlado",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/testerrorasincrono2": {
      "get": {
        "tags": [
          "Test Error"
        ],
        "summary": "Desencadena un error asincrono con promesas then/catch: NO controlado",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/testerrorasincrono3": {
      "get": {
        "tags": [
          "Test Error"
        ],
        "summary": "Desencadena un error asincrono con promesas async/await: Controlado",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/testeventloop": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "peticion",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/jsonexterno": {
      "get": {
        "tags": [
          "Json externo"
        ],
        "summary": "Obtiene un json desde otro sitio, a traves del servidor",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/token": {
      "post": {
        "tags": [
          "Seguridad"
        ],
        "summary": "refresh token",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Seguridad"
        ],
        "summary": "Login de usuarios: admin:123(rol administrador), juan:123(rol miembro)",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario": {
                  "example": "any"
                },
                "clave": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "Seguridad"
        ],
        "summary": "Logout: invalida el refresh token (no invalida el token actual!!!)",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "ArticulosFamilias": {
      "type": "object",
      "properties": {
        "IdArticulo": {
          "type": "number",
          "example": 1
        },
        "Nombre": {
          "type": "string",
          "example": "Accesorios"
        }
      }
    },
    "ArticulosFamiliasMock": {
      "type": "object",
      "properties": {
        "IdArticulo": {
          "type": "number",
          "example": 1
        },
        "Nombre": {
          "type": "string",
          "example": "Accesorios"
        }
      }
    },
    "Articulos": {
      "type": "object",
      "properties": {
        "IdArticulo": {
          "type": "number",
          "example": 143
        },
        "Nombre": {
          "type": "string",
          "example": "Aire acondicionado lg 3000 fc h126tnw0"
        },
        "Precio": {
          "type": "number",
          "example": 7499
        },
        "CodigoDeBarra": {
          "type": "string",
          "example": "0779808338858"
        },
        "IdArticuloFamilia": {
          "type": "number",
          "example": 7
        },
        "Stock": {
          "type": "number",
          "example": 441
        },
        "FechaAlta": {
          "type": "string",
          "example": "2017-01-09T00:00:00"
        },
        "Activo": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}